+++ DRN.Test.Unit/Tests/Framework/Utils/Ids/SourceKnownEntityIdUtilsTests.cs
@@ -472,7 +472,7 @@ public class SourceKnownEntityIdUtilsTests
 
 
     [EntityType(200)]
-    class XEntity(long id) : SourceKnownEntity(id);
+    internal class XEntity(long id) : SourceKnownEntity(id);
     [EntityType(201)]
-    class YEntity(long id) : SourceKnownEntity(id);
+    internal class YEntity(long id) : SourceKnownEntity(id);
 }